:+1: splited mockSource index file
