+++ src/mockSources/mediaQuery.ts
@@ -1,0 +1,10 @@
+import { Stream } from "xstream";
+import { MediaQueryMatch } from "../drivers/mediaQueryDriver";
+
+export interface MockMediaQuerySource {
+  matches$: Stream<MediaQueryMatch>;
+}
+
+export const mockMediaQuerySource = (
+  matches$: Stream<MediaQueryMatch>
+): MockMediaQuerySource => ({ matches$ });
